requirements.txt
fastapi
uvicorn[standard]

langchain
langchain-community
langchain-core
langchain-groq

chromadb

sentence-transformers
PyPDF

python-dotenv

pydantic
requests

aiofiles
streamlit
python-multipart